proxy/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
requests==2.28.2
pydantic==1.10.7
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
ipaddress==1.0.23
prometheus-client==0.16.0
python-dotenv==1.0.0